+++ callbacks/participants.py
@@ -1,7 +1,8 @@
 from telegram import Update
 from telegram.ext import CallbackContext
-from db.queries import find_row_by_chat_id
-from utils.behaviours import print_summary
+from behaviours.print_new_summary import print_new_summary
+from db.queries import find_row_by_chat_id, find_all_info_by_chat_id
+from utils.utils import format_summary
 
 def participants(update: Update, context: CallbackContext):
     chat_id = update.message.chat_id
@@ -11,4 +12,7 @@ def participants(update: Update, context: CallbackContext):
         answer = "Prima di iniziare con le danze, avvia una partita, per farlo usa /start"
         context.bot.send_message(chat_id=update.effective_chat.id, parse_mode='markdown', text=answer)
     else:
-        print_summary(chat_id, False, True, update, context)
+        players, day, time, target, default_message, pitch, teams, bot_last_message_id = find_all_info_by_chat_id(
+            chat_id)
+        current_situation = format_summary(players, day, time, target, default_message, pitch)
+        print_new_summary(current_situation, update, context)
