+++ callbacks/set_number.py
@@ -1,8 +1,12 @@
 from telegram import Update
 from telegram.ext import CallbackContext
-from db.queries import find_all_info_by_chat_id, update_teams_on_db, update_target_on_db
-from utils.behaviours import get_sender_name, filter_maybe_placeholders, remove_job_if_exists, print_summary
 from telegram.utils.helpers import escape_markdown
+from behaviours.edit_summary import edit_summary
+from behaviours.print_new_summary import print_new_summary
+from behaviours.remove_job_if_exists import remove_job_if_exists
+from behaviours.trigger_payment_reminder import trigger_payment_reminder
+from db.queries import find_all_info_by_chat_id, update_teams_on_db, update_target_on_db
+from utils.utils import get_sender_name, filter_maybe_placeholders, format_summary
 
 def set_number(update: Update, context: CallbackContext):
     chat_id = update.message.chat_id
@@ -42,7 +46,18 @@ def set_number(update: Update, context: CallbackContext):
                         update_target_on_db(chat_id, choosen_number)
                         answer = "Ok, " + sender + "! Ho impostato il numero di partecipanti a " + str(choosen_number)
                         reached_target = players and participants_num == choosen_number
-                        print_summary(chat_id, reached_target, False, update, context)
+                        players, day, time, target, default_message, pitch, teams, bot_last_message_id = find_all_info_by_chat_id(
+                            chat_id)
+                        current_situation = format_summary(players, day, time, target, default_message, pitch)
+                        if bot_last_message_id is None:
+                            print_new_summary(chat_id, current_situation, update, context)
+                        else:
+                            edit_summary(current_situation, bot_last_message_id, update, context)
+
+                        if reached_target:
+                            trigger_payment_reminder(update, context, day, time)
+                            context.bot.send_message(chat_id=update.effective_chat.id, parse_mode='markdown',
+                                                     text="🚀 *SI GIOCA* 🚀 facciamo le squadre? /teams 😎")
                 else:
                     answer = sender + ", quello ti sembra un numero? 😂"
