+++ callbacks/set_pitch.py
@@ -1,8 +1,10 @@
 from telegram import Update
 from telegram.ext import CallbackContext
-from db.queries import find_row_by_chat_id, update_pitch_on_db
-from utils.behaviours import flatten_args, get_sender_name, print_summary
 from telegram.utils.helpers import escape_markdown
+from behaviours.edit_summary import edit_summary
+from behaviours.print_new_summary import print_new_summary
+from db.queries import find_row_by_chat_id, update_pitch_on_db, find_all_info_by_chat_id
+from utils.utils import flatten_args, get_sender_name, format_summary
 
 def set_pitch(update: Update, context: CallbackContext):
     chat_id = update.message.chat_id
@@ -21,4 +23,10 @@ def set_pitch(update: Update, context: CallbackContext):
             sender = "@" + get_sender_name(update)
             answer = "Ok, " + sender + "! Ho aggiornato il campo!"
             context.bot.send_message(chat_id=update.effective_chat.id, parse_mode='markdown', text=escape_markdown(answer))
-            print_summary(chat_id, False, False, update, context)
+            players, day, time, target, default_message, pitch, teams, bot_last_message_id = find_all_info_by_chat_id(
+                chat_id)
+            current_situation = format_summary(players, day, time, target, default_message, pitch)
+            if bot_last_message_id is None:
+                print_new_summary(chat_id, current_situation, update, context)
+            else:
+                edit_summary(current_situation, bot_last_message_id, update, context)
